+++ Hazel/include/Renderer/IndexBuffer.h
@@ -9,9 +9,9 @@ namespace Hazel {
       virtual void Bind() const = 0;
       virtual void Unbind() const = 0;
 
-      virtual size_t GetCount() const = 0;
+      virtual uint32_t GetCount() const = 0;
 
-      static IndexBuffer* Create(const unsigned int* indices, const size_t count);
+      static IndexBuffer* Create(const uint32_t* indices, const uint32_t count);
 
    };
 }
