+++ Hazel/src/Renderer/VertexBuffer.cpp
@@ -4,7 +4,7 @@
 #include "Platform/OpenGL/OpenGlVertexBuffer.h"
 
 namespace Hazel {
-   VertexBuffer* VertexBuffer::Create(const float* vertices, const size_t count) {
+   VertexBuffer* VertexBuffer::Create(const float* vertices, const uint32_t count) {
       if (Renderer::GetRendererAPI() == RendererAPI::OpenGL) {
          return new OpenGLVertexBuffer(vertices, count);
       }
